code managing process context and populating process context seperated
