+++ .gitignore
@@ -2,3 +2,4 @@
 cscope.*
 *.o
 *.skel.h
+*.swp
