+++ kalbur/rule_engine/context/Makefile
@@ -2,12 +2,20 @@ ROOT_DIR := ../../..
 
 include $(ROOT_DIR)/Makefile.inc
 
-CONTEXT := context
+MANAGER := context_manager
+POPULATE := populate
 
-$(CONTEXT).o: $(CONTEXT).h $(CONTEXT).c
+all: $(MANAGER).o $(POPULATE).o
+
+$(MANAGER).o: $(MANAGER).h $(MANAGER).c
 	$(call msg,CC,$@)
-	$(Q)$(CC) $(CFLAGS) $(INCLUDES) -c $(CONTEXT).c -o $@
+	$(Q)$(CC) $(CFLAGS) $(INCLUDES) -c $(MANAGER).c -o $@
 
+$(POPULATE).o: $(POPULATE).h $(POPULATE).c
+	$(call msg,CC,$@)
+	$(Q)$(CC) $(CFLAGS) $(INCLUDES) -c $(POPULATE).c -o $@
 
 clean:
 	$(Q)rm -f *.o
+
+.PHONY: all clean
