+++ kalbur/rule_engine/context/populate.h
@@ -1,0 +1,7 @@
+#ifndef POPULATE_H
+#define POPULATE_H
+#include "context_manager.h"
+
+int add_event_context(struct process_context *ctx, struct message_state *ms);
+
+#endif // POPULATE_H
